Removed ali2xyz.py script from repo, ali2xyz.py is now created on the fly by the best_enantiomers.sh script. This helps get around issues with the incorrect python being called, which resulted in errors with the EMAN2 module not being found, which is required by the ali2xyz.py script.
superdenss updated to convert final_avg_ali2ref.hdf file to MRC
format and move it to the main folder, rather than buried in the
spt_avg_01 directory.

Added the fsc2res.py script to calculate resolution from final
fsc_0.txt and plot the fsc curve to a .png file. The
resolution is now printed to the screen at the end of the superdenss
script.

Version bump to 1.0.3